Add an Avro format conversion to the command line app
This is a major change largely because it required changing the
SonarDataRecord enum struct variants into tuple variants with structs
for each of the data types. This way we can serialize and deserialize
the data structs individually, which fits a little better with the
Avro data model. Having done this, however, the Avro file creation is
quite simple: we just need to create an iterator over whatever sonar
data we want to store, and then we use the writer.extend_ser method to
write each record to the file.