Allow different sonar data types
This is one way that one can allow different data types for the sonar
data. We use the bytes_per_sample and binrw's pre_assert to dispatch
to different variants of the SonarData enum. A remaining challenge is
to retrieve the bytes_per_sample flag, which is stored in the ChanInfo
structure corresponding to the data channel.