Draw the WaterfallPlot as a widget
There is still a bit of abstraction leaking: the renderer manages the
texture loading from the app, but that should be relatively
straightforward to address. The waterfall widget is also drawn under
all of the vello widgets, which seems somewhat unavoidable unless we
can plug the sonar textures into the vello pipeline. Currently this
means you have to be careful with what you place and where to ensure
that it isn't covered up.

It is also not aligned with the PingPlot widget.