Ensure that the port and starboard channels are aligned
This relies on filtering the ping iterators so that the port channel
and the starboard channel each have the same timestamps. This won't
work if port/starboard pings do not have equal timestamps, which may
not always be true. It will also produce incorrect results if there
are multiple sonar subsystems in the file.