Make the frequency a u32
This may not be particularly accurate, but we are predominantly using
the frequency to identify channels in multi-frequency systems, so it
is more important that it be a unique, Hash/Eq, value than an exact
floating point representation of the frequency.

Also some of our JSF data doesn't have a floating point
mixer_frequency set, but does have the integer start and end
frequencies.